chore: move to non-index for agent and voltagent + add lightweight public API test
