+++ packages/core/src/agent/agent.spec.ts
@@ -6,7 +6,7 @@ import type { Memory, MemoryMessage } from "../memory/types";
 import { AgentRegistry } from "../server/registry";
 import { createTool } from "../tool";
 import { createAsyncIterableStream } from "../utils/async-iterable-stream";
-import { Agent } from "./index";
+import { Agent } from "./agent";
 import type {
   BaseMessage,
   BaseTool,
@@ -18,13 +18,13 @@ import type {
   StepWithContent,
 } from "./providers";
 
-// @ts-ignore - To simplify test types
-import type { AgentHistoryEntry } from "../agent/history";
 import type { NewTimelineEvent } from "../events/types";
 import type { BaseRetriever } from "../retriever/retriever";
 import type { VoltAgentExporter } from "../telemetry/exporter";
 import type { Tool, Toolkit } from "../tool";
 import { streamEventForwarder } from "../utils/streams/stream-event-forwarder";
+// @ts-ignore - To simplify test types
+import type { AgentHistoryEntry } from "./history";
 import { HistoryManager } from "./history";
 import { createHooks } from "./hooks";
 import type { AgentStatus, OperationContext, ToolExecutionContext } from "./types";
